fix: never mind, github actions docs are atrocious
seriously, was this designed to waste corporate money?
what am i misunderstanding about these docs?

https://docs.github.com/en/actions/writing-workflows/workflow-syntax-for-github-actions#onpushpull_requestpull_request_targetpathspaths-ignore

there is nothing ambiguous about these statements:

> If you define both branches/branches-ignore and paths/paths-ignore,
the workflow will only run when both filters are satisfied.

> The paths and paths-ignore keywords accept glob patterns that use the
* and ** wildcard characters to match more than one path name.

unless i am fundamentally misunderstanding something, i believe i have
specified a config that means "run this action when a push to the main
branch changes any file in the src/ or static/ directory", and i am
pushing a change in the src/ directory to the main branch.

the only thing i can think of is that maybe the paths feature requires
changes to be made in all of the specified paths? but it explicitly
says:

> If at least one path matches a pattern in the paths filter, the
workflow runs.

seriously, what is the deal here???